feat: traits for hooks and resources
