+++ src/Traits/Resources.php
@@ -1,0 +1,90 @@
+<?php
+
+namespace Utopia\Traits;
+
+use Exception;
+
+trait Resources
+{
+    /**
+     * @var array
+     */
+    protected array $resources = [
+        'error' => null,
+    ];
+
+    /**
+     * @var array
+     */
+    protected static array $resourcesCallbacks = [];
+
+    /**
+     * If a resource has been created return it, otherwise create it and then return it
+     *
+     * @param string  $name
+     * @param bool  $fresh
+     * @return mixed
+     *
+     * @throws Exception
+     */
+    public function getResource(string $name, bool $fresh = false): mixed
+    {
+        if ($name === 'utopia') {
+            return $this;
+        }
+
+        if (!\array_key_exists($name, $this->resources) || $fresh || self::$resourcesCallbacks[$name]['reset']) {
+            if (!\array_key_exists($name, self::$resourcesCallbacks)) {
+                throw new Exception('Failed to find resource: "' . $name . '"');
+            }
+
+            $this->resources[$name] = \call_user_func_array(
+                self::$resourcesCallbacks[$name]['callback'],
+                $this->getResources(self::$resourcesCallbacks[$name]['injections'])
+            );
+        }
+
+        self::$resourcesCallbacks[$name]['reset'] = false;
+
+        return $this->resources[$name];
+    }
+
+    /**
+     * Get Resources By List
+     *
+     * @param array  $list
+     * @return array
+     */
+    public function getResources(array $list): array
+    {
+        $resources = [];
+
+        foreach ($list as $name) {
+            $resources[$name] = $this->getResource($name);
+        }
+
+        return $resources;
+    }
+
+    /**
+     * Set a new resource callback
+     *
+     * @param string  $name
+     * @param callable  $callback
+     * @param array  $injections
+     * @return void
+     *
+     * @throws Exception
+     */
+    public static function setResource(string $name, callable $callback, array $injections = []): void
+    {
+        if ($name === 'utopia') {
+            throw new Exception("'utopia' is a reserved keyword.", 500);
+        }
+        self::$resourcesCallbacks[$name] = [
+            'callback' => $callback,
+            'injections' => $injections,
+            'reset' => true
+        ];
+    }
+}
